fix(nes): fix uint16_t wrap in add_register_range causing infinite loop on mapper 1 ROMs
Loop variable uint16_t wraps from 0xFFFF to 0x0000 making i <= 0xFFFF
always true. Use uint32_t instead. Affects any mapper calling
add_register_range(0x8000, 0xFFFF, ...).

+++ src/nes/base_mapper.cpp
@@ -51,7 +51,7 @@ void BaseMapper::write_register(uint16_t addr, uint8_t value) {
 }
 
 void BaseMapper::add_register_range(uint16_t start, uint16_t end, MemoryOperation operation) {
-    for (uint16_t i = start; i <= end; i++) {
+    for (uint32_t i = start; i <= end; i++) {
         if (operation == MemoryOperation::READ || operation == MemoryOperation::ANY)
             is_read_register_addr_[i] = true;
         if (operation == MemoryOperation::WRITE || operation == MemoryOperation::ANY)
